# API_documentation/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/mastertables/departments": {
      "get": {
        "operationId": "DepartmentController_getDepartments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of departments available"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all departments",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "DepartmentController_createDepartment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create Department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new department",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/departments/{id}": {
      "get": {
        "operationId": "DepartmentController_getDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Here is the department"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a department",
        "tags": [
          "Mastertables"
        ]
      },
      "put": {
        "operationId": "DepartmentController_updateDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update department",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a current department information",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/companies": {
      "get": {
        "operationId": "CompanyController_getCompanies",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of companies retrieved"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all companies",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "CompanyController_createCompany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create company",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new company",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/companies/{id}": {
      "get": {
        "operationId": "CompanyController_getCompany",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Here is the company"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a company",
        "tags": [
          "Mastertables"
        ]
      },
      "put": {
        "operationId": "CompanyController_updateCompany",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update company",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Company updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a current company information",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/divisions": {
      "get": {
        "operationId": "DivisionController_getDivisions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of divisions retrieved"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all divisions",
        "tags": [
          "Mastertables"
        ]
      },
      "post": {
        "operationId": "DivisionController_createDivision",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Payload to create Division",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDivisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Division created successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "409": {
            "description": "Conflict - Resource already exist or duplicate entry"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new division",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/divisions/{id}": {
      "get": {
        "operationId": "DivisionController_getDivision",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Here is the division"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a division",
        "tags": [
          "Mastertables"
        ]
      }
    },
    "/mastertables/divisions/{divisionId}": {
      "put": {
        "operationId": "DivisionController_updateDivision",
        "parameters": [
          {
            "name": "divisionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Payload to update division",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDivisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Division updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid or missing input or parameters"
          },
          "401": {
            "description": "Unauthorized - You do not have access to this resource"
          },
          "403": {
            "description": "Forbidden - You do not have access to this resource"
          },
          "404": {
            "description": "Not Found - The requested resource was not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a current division information",
        "tags": [
          "Mastertables"
        ]
      }
    },
...
  },
  "info": {
    "title": "ABAS v3 API",
    "description": "API documentation for ABAS v3 project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      },
      "security_clearance": {
        "type": "apiKey",
        "name": "Security-Clearance",
        "in": "header",
        "description": "Required security clearance level (documentational only)."
      }
    },
    "schemas": {
...
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "department_name": {
            "type": "string",
            "example": "Human Resources",
            "description": "The name of the department"
          },
          "sorting": {
            "type": "number",
            "example": 2,
            "description": "Sorting number of the department"
          },
          "division": {
            "type": "number",
            "example": "corporate services = 1, asset management = 2, marketing and operations = 3, cebu air inc = 4",
            "description": "The Division where the department belongs to"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "department_name",
          "sorting",
          "division",
          "status"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "department_name": {
            "type": "string",
            "example": "New Department name",
            "description": "If you want to update the Department name"
          },
          "sorting": {
            "type": "number",
            "example": 2,
            "description": "Sorting number of the department"
          },
          "division": {
            "type": "number",
            "example": "corporate services = 1, asset management = 2, marketing and operations = 3, cebu air inc = 3",
            "description": "The Division where the department belongs to"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "department_name",
          "sorting",
          "division",
          "status"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Avega Bros. Integrated Shipping Corp",
            "description": "Name of the company"
          },
          "address": {
            "type": "string",
            "example": "Cebu City, Cebu",
            "description": "Address or location of the company"
          },
          "telephone_no": {
            "type": "string",
            "example": "011-63-45-123-4567",
            "description": "Telephone No. of the company"
          },
          "fax_no": {
            "type": "string",
            "example": "011-63-2-1234567",
            "description": "Fax No. of the company"
          },
          "company_tin": {
            "type": "string",
            "example": "11-M1115-1234",
            "description": "Registered Tax Identification No. of the company"
          },
          "is_top_20000": {
            "type": "number",
            "example": "yes or no",
            "description": "yes = 1, no = 0"
          },
          "abbreviation": {
            "type": "string",
            "example": "ABISC",
            "description": "Short name or abbreviation of the company"
          },
          "stat": {
            "type": "number",
            "example": "active or inactive",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "name",
          "address",
          "telephone_no",
          "fax_no",
          "company_tin",
          "is_top_20000",
          "abbreviation",
          "stat"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Avega Bros. Integrated Shipping Corp",
            "description": "Update current company name"
          },
          "address": {
            "type": "string",
            "example": "Cebu City, Cebu",
            "description": "Update current company location"
          },
          "telephone_no": {
            "type": "string",
            "example": "011-63-45-123-4567",
            "description": "Update current company telephone no."
          },
          "fax_no": {
            "type": "string",
            "example": "011-63-2-1234567",
            "description": "Update current company fax no."
          },
          "company_tin": {
            "type": "string",
            "example": "11-M1115-1234",
            "description": "Update current company tin no."
          },
          "is_top_20000": {
            "type": "number",
            "example": "yes or no",
            "description": "Update current company is top 20000?"
          },
          "abbreviation": {
            "type": "string",
            "example": "ABISC",
            "description": "Update current company abbreviation"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "Update current company status"
          }
        },
        "required": [
          "name",
          "address",
          "telephone_no",
          "fax_no",
          "company_tin",
          "is_top_20000",
          "abbreviation",
          "status"
        ]
      },
      "CreateDivisionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Cebu Air Inc",
            "description": "The name of the division"
          },
          "division_head_id": {
            "type": "number",
            "example": 3,
            "description": "ID number of the division head it belongs to"
          },
          "status": {
            "type": "number",
            "example": "active",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "name",
          "division_head_id",
          "status"
        ]
      },
      "UpdateDivisionDto": {
        "type": "object",
        "properties": {
          "division_id": {
            "type": "number",
            "example": 1,
            "description": "ID of the division you want to update"
          },
          "division_name": {
            "type": "string",
            "example": "New Division Name",
            "description": "If you want to update the Division name"
          },
          "status": {
            "type": "number",
            "example": "active or inactive",
            "description": "active = 1, inactive = 0"
          }
        },
        "required": [
          "division_id",
          "division_name",
          "status"
        ]
      },
...
    }
  }
}